refactor: decouple engine errors from internal/db via adapter layer
